+++ stable_baselines3/common/vec_env/__init__.py
@@ -1,15 +1,9 @@
 # flake8: noqa F401
 import typing
 from copy import deepcopy
-from typing import Optional, Union
+from typing import Optional, Type, Union
 
-from stable_baselines3.common.vec_env.base_vec_env import (
-    AlreadySteppingError,
-    CloudpickleWrapper,
-    NotSteppingError,
-    VecEnv,
-    VecEnvWrapper,
-)
+from stable_baselines3.common.vec_env.base_vec_env import CloudpickleWrapper, VecEnv, VecEnvWrapper
 from stable_baselines3.common.vec_env.dummy_vec_env import DummyVecEnv
 from stable_baselines3.common.vec_env.subproc_vec_env import SubprocVecEnv
 from stable_baselines3.common.vec_env.vec_check_nan import VecCheckNan
@@ -23,17 +17,28 @@ if typing.TYPE_CHECKING:
     from stable_baselines3.common.type_aliases import GymEnv
 
 
+def unwrap_vec_wrapper(env: Union["GymEnv", VecEnv], vec_wrapper_class: Type[VecEnvWrapper]) -> Optional[VecEnvWrapper]:
+    """
+    Retrieve a ``VecEnvWrapper`` object by recursively searching.
+
+    :param env: (gym.Env)
+    :param vec_wrapper_class: (VecEnvWrapper)
+    :return: (VecEnvWrapper)
+    """
+    env_tmp = env
+    while isinstance(env_tmp, VecEnvWrapper):
+        if isinstance(env_tmp, vec_wrapper_class):
+            return env_tmp
+        env_tmp = env_tmp.venv
+    return None
+
+
 def unwrap_vec_normalize(env: Union["GymEnv", VecEnv]) -> Optional[VecNormalize]:
     """
     :param env: (gym.Env)
     :return: (VecNormalize)
     """
-    env_tmp = env
-    while isinstance(env_tmp, VecEnvWrapper):
-        if isinstance(env_tmp, VecNormalize):
-            return env_tmp
-        env_tmp = env_tmp.venv
-    return None
+    return unwrap_vec_wrapper(env, VecNormalize)  # pytype:disable=bad-return-type
 
 
 # Define here to avoid circular import