+++ docs/misc/changelog.rst
@@ -15,16 +15,19 @@ New Features:
 
 Bug Fixes:
 ^^^^^^^^^^
+- Added ``unwrap_vec_wrapper()`` to ``common.vec_env`` to extract ``VecEnvWrapper`` if needed
+- Fixed a bug where the environment was reset twice when using ``evaluate_policy``
 
 Deprecations:
 ^^^^^^^^^^^^^
 
 Others:
 ^^^^^^^
+- Improve typing coverage of the ``VecEnv``
+- Removed ``AlreadySteppingError`` and ``NotSteppingError`` that were not used
 
 Documentation:
 ^^^^^^^^^^^^^^
-
 
 Pre-Release 0.8.0 (2020-08-03)
 ------------------------------